+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorAdditionalPropertiesTest.kt
@@ -376,16 +376,17 @@ class CodeGeneratorAdditionalPropertiesTest {
         expect(resultFile("TestApTrueConst")) { outputDetails.output() }
     }
 
-//    @Test fun `should generate code for experiment`() {
-//        val input = File("src/test/resources/test-xxx.schema.json")
-//        val outputDetails = OutputDetails(TargetFileName("TestXxx", "kt", packageDirs))
-//        CodeGenerator().apply {
-//            configure(File("src/test/resources/config/ap-config.json"))
-//            outputResolver = outputCapture(outputDetails)
-//            generate(input)
-//        }
-//        expect("Dummy") { outputDetails.output() }
-//    }
+    @Test fun `should generate code for extensible enum`() {
+        val input = File("src/test/resources/test-extensible-enum.schema.json")
+        val config = File("src/test/resources/config/extensible-enum-config.json")
+        val outputDetails = OutputDetails(TargetFileName("TestExtensibleEnum", "kt", packageDirs))
+        CodeGenerator().apply {
+            configure(config)
+            outputResolver = outputCapture(outputDetails)
+            generate(input)
+        }
+        expect(resultFile("TestExtensibleEnum")) { outputDetails.output() }
+    }
 
     // TODO - more tests of nested classes?